+++ pylint_nautobot/deprecated_status_model.py
@@ -15,8 +15,7 @@ class NautobotDeprecatedStatusModelChecker(BaseChecker):
         "E4292": (
             "Inherits from the deprecated StatusModel instead of declaring status on the model explicitly with StatusField",
             "nb-status-field-instead-of-status-model",
-            # TBD: Link is not accessible
-            "https://docs.nautobot.com/projects/core/en/next/models/extras/status/#status-internals",
+            "Reference: https://docs.nautobot.com/projects/core/en/next/user-guide/platform-functionality/status/#status-internals",
         ),
     }