+++ docs/release_notes.md
@@ -10,6 +10,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1] - 2023-09-01
+
+### Changed
+
+- Updates CODEOWNERS (#41)
+- Improves rule help messages (#49)
+- Fix scoping on string field blank/null checker (#49)
+- Parametrize the base class checker by the Nautobot version (#44)
+
 ## [0.2.0] - 2023-06-28
 
 ### Added